Refactor for OOP
